requirements.txt
aiogram==3.13.1
fastapi==0.115.6
uvicorn[standard]==0.32.1
asyncpg==0.30.0
python-dotenv==1.0.1
google-api-python-client==2.154.0
google-auth==2.37.0
